wp-comment-sentiment/requirements.txt
transformers==4.34.1
kafka-python
torch
emoji==0.6.0
